Separate into entrypoints
