+++ docs/docs/modules/indexes/retrievers/self_query/index.mdx
