+++ langchain/package.json
@@ -328,9 +328,15 @@
     "retrievers/self_query.cjs",
     "retrievers/self_query.js",
     "retrievers/self_query.d.ts",
+    "retrievers/self_query/chroma.cjs",
+    "retrievers/self_query/chroma.js",
+    "retrievers/self_query/chroma.d.ts",
     "retrievers/self_query/functional.cjs",
     "retrievers/self_query/functional.js",
     "retrievers/self_query/functional.d.ts",
+    "retrievers/self_query/pinecone.cjs",
+    "retrievers/self_query/pinecone.js",
+    "retrievers/self_query/pinecone.d.ts",
     "retrievers/self_query/supabase.cjs",
     "retrievers/self_query/supabase.js",
     "retrievers/self_query/supabase.d.ts",
@@ -1255,10 +1261,20 @@
       "import": "./retrievers/self_query.js",
       "require": "./retrievers/self_query.cjs"
     },
+    "./retrievers/self_query/chroma": {
+      "types": "./retrievers/self_query/chroma.d.ts",
+      "import": "./retrievers/self_query/chroma.js",
+      "require": "./retrievers/self_query/chroma.cjs"
+    },
     "./retrievers/self_query/functional": {
       "types": "./retrievers/self_query/functional.d.ts",
       "import": "./retrievers/self_query/functional.js",
       "require": "./retrievers/self_query/functional.cjs"
+    },
+    "./retrievers/self_query/pinecone": {
+      "types": "./retrievers/self_query/pinecone.d.ts",
+      "import": "./retrievers/self_query/pinecone.js",
+      "require": "./retrievers/self_query/pinecone.cjs"
     },
     "./retrievers/self_query/supabase": {
       "types": "./retrievers/self_query/supabase.d.ts",
