+++ langchain/scripts/create-entrypoints.js
@@ -134,7 +134,9 @@ const entrypoints = {
     "retrievers/document_compressors/chain_extract",
   "retrievers/hyde": "retrievers/hyde",
   "retrievers/self_query": "retrievers/self_query/index",
+  "retrievers/self_query/chroma": "retrievers/self_query/chroma",
   "retrievers/self_query/functional": "retrievers/self_query/functional",
+  "retrievers/self_query/pinecone": "retrievers/self_query/pinecone",
   "retrievers/self_query/supabase": "retrievers/self_query/supabase",
   "retrievers/vespa": "retrievers/vespa",
   // cache
@@ -234,7 +236,9 @@ const requiresOptionalDependency = [
   "retrievers/zep",
   "retrievers/metal",
   "retrievers/self_query",
+  "retrievers/self_query/chroma",
   "retrievers/self_query/functional",
+  "retrievers/self_query/pinecone",
   "retrievers/self_query/supabase",
   "output_parsers/expression",
   "chains/query_constructor",
