+++ langchain/tsconfig.json
@@ -132,7 +132,9 @@
       "src/retrievers/document_compressors/chain_extract.ts",
       "src/retrievers/hyde.ts",
       "src/retrievers/self_query/index.ts",
+      "src/retrievers/self_query/chroma.ts",
       "src/retrievers/self_query/functional.ts",
+      "src/retrievers/self_query/pinecone.ts",
       "src/retrievers/self_query/supabase.ts",
       "src/retrievers/vespa.ts",
       "src/cache/index.ts",
